[Refactor]: GroupController 에 그룹 가입과 관련된 API GroupJoinController 로 이전
